Feature/mysql integration

+++ CadastroProdutoMySQL/Modelos/Produto.cs
@@ -5,7 +5,7 @@
         // Atributos da classe
         public int Id { get; set; }
         public string Nome { get; set; }
-        public double Preco { get; set; }
+        public decimal Preco { get; set; }
 
         // Construtor vazio
         public Produto()
@@ -13,7 +13,7 @@
         }
 
         // Construtor com parametros
-        public Produto(int id, string nome, double preco)
+        public Produto(int id, string nome, decimal preco)
         {
             Id = id;
             Nome = nome;
